pyproject.toml
[project]
name = "spacialarm"
version = "1.0.0"
description = "Visualizador 3D del Brazo Robótico SpacialArm"
authors = [
    {name = "SpacialArm Team", email = "team@spacialarm.com"}
]
readme = "Readme.md"
requires-python = ">=3.12"
dependencies = [
    "PyOpenGL>=3.1.7",
    "PyOpenGL_accelerate>=3.1.7", 
    "PyQt5>=5.15.10",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = [
    "*.py",
    "window/*.py",
    "window/*.ui",
    "*.md",
    "requirements.txt",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0", 
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
